refactor(studio-integration): Remove NamePromise
Something seems to be wrong having the Promise for content
name as extra class. It behaves weird, the constructor got
called twice, once with an invalid argument.

A simple reproducer at TypeScript Playground: https://t1p.de/o15m

Decided to use the passive approach, thus to build the Promise
inline again in MockContentDisplayService.
